registro_de_nota/registro_de_notas/requirements.txt
certifi==2018.4.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.0.5
django-cors-middleware==1.3.1
django-filter==1.1.0
django-oauth-toolkit==1.2.0
django-rest-swagger==2.2.0
djangorestframework==3.8.2
djangorestframework-simplejwt==3.2.3
drf-nested-routers==0.90.2
idna==2.7
itypes==1.1.0
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.0
oauthlib==2.1.0
openapi-codec==1.3.2
Pygments==2.2.0
PyJWT==1.6.4
pytz==2018.4
requests==2.19.1
simplejson==3.16.0
uritemplate==3.0.0
urllib3==1.23
